requirements.txt
azure-identity
msgraph-sdk
azure-cosmos